requirements.txt
autopep8==1.4
numpy==1.15.1
python-midi==0.2.4
tensorflow==1.10.1
